move hexwidget into it's own package
The original hexwidget PR started before fyne-x really had any
established conventions, so I put it in it's own package. Since then, it
seems we've settled on keeping things all together in the widget
package.
