--- widget/hexwidget/README.md
@@ -1,16 +1,0 @@
-# Hex Widget
-
-This widget implements a skeuomorphic "7 segment" hexadecimal display.
-
-For an example of usage, see [the hexwidget demo](../../cmd/hexwidget_demo).
-
-## Screenshots
-
-![](hexwidget_00abcdef.png)
-
-![](hexwidget_12345678.png)
-
-## Future Work
-
-* Add a widget that combines multiple hex displays into a single *n*-bit number
-  (where *n* is a multiple of 4).
